Adding Toaster

+++ frontend/src/pages/ApplicationPage.tsx
@@ -1,6 +1,7 @@
 // useLoaderData is the mehcanism to fetch the record and allows to use the function elsewhere
 import { useLoaderData, useParams, useNavigate } from "react-router-dom";
 import { Link } from "react-router-dom";
+import { toast } from "react-toastify";
 
 async function applicationLoader({ params }) {
   try {
@@ -21,21 +22,26 @@ function ApplicationPage() {
   const navigate = useNavigate(); // Add this hook
 
   const deleteApplication = async () => {
-    //console.log(params);
-    try {
-      const res = await fetch(`/api/application/${params.applicationID}`, {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-      if (res.status != 201) {
+    const confirm: boolean = window.confirm(
+      "Are you sure you want to delete the record?"
+    );
+    if (confirm) {
+      try {
+        const res = await fetch(`/api/application/${params.applicationID}`, {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+          },
+        });
+        if (res.status != 201) {
+          return { success: false };
+        }
+        toast.success("Application deleted successfully");
+        navigate("/applications");
+      } catch (error) {
+        toast.error("Unable to delete application");
         return { success: false };
       }
-      console.log("REdirect:" + res.status);
-      navigate("/applications");
-    } catch (error) {
-      return { success: false };
     }
   };
